+ This is another major overhaul of the RECORD/REPLAY Code. The biggest change   is having single log entry per event. Here is a list of major changes:   - One log entry per synchronization event.   - As a side-effect, the XXX_return events don't exist anymore.   - Log patching not required now.   - Log annotation is done on-the-fly at the time of record itself.   - Fake replay of pthread mutexes, condition variables, and rwlocks. + This changeset is an ongoing work, but its getting very close to become   stable.

+++ trunk/dmtcp/src/virtualpidtable.cpp
@@ -485,7 +485,7 @@ void dmtcp::VirtualPidTable::serializePidMap ( jalib::JBinarySerializer& o )
   }
   else
   {
-    for (int i = 0; i < numMaps; i++) {
+    for (size_t i = 0; i < numMaps; i++) {
       serializePidMapEntry ( o, originalPid, currentPid );
       _pidMapTable[originalPid] = currentPid;
     }
